third_party/WebKit/LayoutTests/inspector/documentation/documentation-url-provider.html

<html>
<head>
<script src="../../http/tests/inspector/inspector-test.js"></script>
<script>

function test()
{
    InspectorTest.registerModule("documentation", true);
    var testCases = [
        "max",
        "insertBefore",
        "eval",
        "stringify",
        "isArray",
        "now",
        "fromCharCode",
        "keys",
        "notExist",
        "indexOf",
        "setDate"
        ];
    var urlFormer = new WebInspector.DocumentationURLProvider();
    for (var i = 0; i < testCases.length; ++i) {
        InspectorTest.addResult("Property " + testCases[i] + ":");
        InspectorTest.addResult(urlFormer.itemPath(testCases[i]));
    }
    InspectorTest.completeTest();
}

</script>
</head>
<body onload="runTest()">
This test verifies URL created by WebInspector.DocumentationURLProvider
</body>
</html>
